refactor(图表): typo
