+++ core/core-frontend/src/store/modules/data-visualization/snapshot.ts
@@ -2,7 +2,7 @@ import { defineStore, storeToRefs } from 'pinia'
 import { store } from '../../index'
 import { dvMainStoreWithOut } from './dvMain'
 import { deepCopy } from '@/utils/utils'
-import { BASE_THEMES } from '@/views/chart/components/editor/util/dataVisualiztion'
+import { BASE_THEMES } from '@/views/chart/components/editor/util/dataVisualization'
 import eventBus from '@/utils/eventBus'
 import { useEmitt } from '@/hooks/web/useEmitt'
 import { useCache } from '@/hooks/web/useCache'
