+++ core/core-frontend/src/views/chart/components/editor/util/dataVisualization.ts
